Refactor of hydrology module names complete, including fortran compile
