+++ pywatershed/hydrology/prms_channel.pyf
@@ -1,10 +1,10 @@
 !    -*- f90 -*-
 ! Note: the context of this file is case sensitive.
 
-python module PRMSChannel_f ! in 
-    interface  ! in :PRMSChannel_f
+python module prms_channel_f ! in 
+    interface  ! in :prms_channel_f
         subroutine calc_muskingum_mann( &
-            nseg,segment_order,to_segment,seg_lateral_inflow,seg_inflow0_in,outflow_ts_in,tsi,ts,c0,c1,c2,seg_upstream_inflow,seg_inflow0,seg_inflow,seg_outflow,inflow_ts,outflow_ts,seg_current_sum) ! in :PRMSChannel_f:PRMSChannel.f90
+            nseg,segment_order,to_segment,seg_lateral_inflow,seg_inflow0_in,outflow_ts_in,tsi,ts,c0,c1,c2,seg_upstream_inflow,seg_inflow0,seg_inflow,seg_outflow,inflow_ts,outflow_ts,seg_current_sum) ! in :prms_channel_f:prms_channel.f90
             integer(kind=4), optional,intent(in),check(shape(segment_order, 0) == nseg),depend(segment_order) :: nseg=shape(segment_order, 0)
             integer(kind=8) dimension(nseg),intent(in) :: segment_order
             integer(kind=8) dimension(nseg),intent(in),depend(nseg) :: to_segment
@@ -25,7 +25,7 @@ python module PRMSChannel_f ! in
             real(kind=8) dimension(nseg),intent(out),depend(nseg) :: seg_current_sum
         end subroutine calc_muskingum_mann
     end interface 
-end python module PRMSChannel_f
+end python module prms_channel_f
 
 ! This file was auto-generated with f2py (version:1.23.3).
 ! See:
