+++ pywatershed/hydrology/prms_groundwater.pyf
@@ -1,9 +1,9 @@
 !    -*- f90 -*-
 ! Note: the context of this file is case sensitive.
 
-python module PRMSGroundwater_f
+python module prms_groundwater_f
     interface
-        subroutine calc_groundwater(nhru,gwarea,soil_to_gw,ssr_to_gw,dprst_seep_hru,gwres_stor_in,gwflow_coef,gwsink_coef,gwres_stor_old,hru_in_to_cf,gwres_stor_out,gwres_flow,gwres_sink,gwres_stor_change,gwres_flow_vol) ! in PRMSGroundwater.f90
+        subroutine calc_groundwater(nhru,gwarea,soil_to_gw,ssr_to_gw,dprst_seep_hru,gwres_stor_in,gwflow_coef,gwsink_coef,gwres_stor_old,hru_in_to_cf,gwres_stor_out,gwres_flow,gwres_sink,gwres_stor_change,gwres_flow_vol) ! in prms_groundwater.f90
             integer, optional,intent(in),check(len(gwarea)>=nhru),depend(gwarea) :: nhru=len(gwarea)
             real(kind=8) dimension(nhru),intent(in) :: gwarea
             real(kind=8) dimension(nhru),intent(in) :: soil_to_gw
@@ -21,7 +21,7 @@ python module PRMSGroundwater_f
             real(kind=8) dimension(nhru),intent(out) :: gwres_flow_vol
         end subroutine calc_groundwater
     end interface
-end python module PRMSGroundwater_f
+end python module prms_groundwater_f
 
 ! This file was auto-generated with f2py (version:1.21.6).
 ! See http://cens.ioc.ee/projects/f2py2e/
